+++ jdaviz/configs/specviz/plugins/unit_conversion/unit_conversion.vue
@@ -62,7 +62,7 @@
       ></v-select>
     </v-row>
 
-    <v-row v-if="has_spectral">
+    <v-row v-if="has_sb">
       <v-text-field
         v-model="sb_unit_selected"
         :label="api_hints_enabled ? 'plg.sb_unit' : 'Surface Brightness Unit'"