+++ jdaviz/core/marks.py
@@ -6,6 +6,7 @@ from bqplot.marks import Lines, Label, Scatter
 from glue.core import HubListener
 from specutils import Spectrum1D
 from jdaviz.utils import _eqv_pixar_sr, _eqv_flux_to_sb_pixel
+from jdaviz.core.validunits import check_if_unit_is_per_solid_angle
 
 from jdaviz.core.events import GlobalDisplayUnitChanged
 from jdaviz.core.events import (SliceToolStateMessage, LineIdentifyMessage,
@@ -125,6 +126,12 @@ class PluginMark:
             unit = self.viewer.state.y_display_unit
         unit = u.Unit(unit)
 
+        # spectrum y-values in viewer have already been converted, don't convert again
+        # if a spectral_y_type is changed, just update the unit
+        if self.yunit is not None and check_if_unit_is_per_solid_angle(self.yunit) != check_if_unit_is_per_solid_angle(unit):  # noqa
+            self.yunit = unit
+            return
+
         if self.yunit is not None and not np.all([s == 0 for s in self.y.shape]):
             if self.viewer.default_class is Spectrum1D:
                 # used to obtain spectral density equivalencies with previous data and units
@@ -149,7 +156,9 @@ class PluginMark:
     def _on_global_display_unit_changed(self, msg):
         if not self.auto_update_units:
             return
-        if self.viewer.__class__.__name__ in ['SpecvizProfileView', 'CubevizProfileView']:
+        if self.viewer.__class__.__name__ in ['SpecvizProfileView',
+                                              'CubevizProfileView',
+                                              'MosvizProfileView']:
             axis_map = {'spectral': 'x', 'spectral_y': 'y'}
         elif self.viewer.__class__.__name__ == 'MosvizProfile2DView':
             axis_map = {'spectral': 'x'}
@@ -157,6 +166,12 @@ class PluginMark:
             return
         axis = axis_map.get(msg.axis, None)
         if axis is not None:
+            scale = self.scales.get(axis, None)
+            # if PluginMark mark is LinearScale(0, 1), prevent it from entering unit conversion
+            # machinery so it maintains it's position in viewer.
+            if isinstance(scale, LinearScale) and (scale.min, scale.max) == (0, 1):
+                return
+
             getattr(self, f'set_{axis}_unit')(msg.unit)
 
     def clear(self):