+++ CHANGES.rst
@@ -21,6 +21,8 @@ Specviz
 Specviz2d
 ^^^^^^^^^
 
+- Implement the Unit Conversion plugin the Specviz2D. [#3253]
+
 API Changes
 -----------